requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
fpdf2==2.7.7
